requirements.txt
pandas
openpyxl
numpy
scikit-learn
fastapi
uvicorn
requests
python-dotenv
fpdf2